modified success to use Pattern1 to render page
